Fix number representation for short (< 8 bytes) numbers
