+++ pkg/ebpf/util/endianess.go
@@ -1,0 +1,26 @@
+package util
+
+import (
+	"encoding/binary"
+	"unsafe"
+)
+
+var nativeEndian binary.ByteOrder
+
+func init() {
+	buf := [2]byte{}
+	*(*uint16)(unsafe.Pointer(&buf[0])) = uint16(0xABCD)
+
+	switch buf {
+	case [2]byte{0xCD, 0xAB}:
+		nativeEndian = binary.LittleEndian
+	case [2]byte{0xAB, 0xCD}:
+		nativeEndian = binary.BigEndian
+	default:
+		panic("Could not determine native endianness.")
+	}
+}
+
+func GetEndian() binary.ByteOrder {
+	return nativeEndian
+}
